requirements.txt
certifi==2024.2.2
charset-normalizer==3.4.0
idna==3.7
python-dotenv==1.0.1
redis==6.1.0
requests==2.31.0
six==1.17.0
spotipy==2.23.0
urllib3==2.2.1
wxPython==4.2.3
